+++ NEWS.md
@@ -1,7 +1,31 @@
 ------------------------------------------------------------------------
-adnuts 1.1.0.9000
+adnuts 1.1.1.9000
 ------------------------------------------------------------------------
+
+* Improve console output for RStudio users. Was broken for NUTS chains and in serial.
+
+* Add new option -verbose which suppresses almost all console output when set to FALSE
+
+* Update demo file, vignette and README in preparation for submission to CRAN
+
+* Add smoother 
+
+------------------------------------------------------------------------
+adnuts 1.1.1 (2021-02-19)
+------------------------------------------------------------------------
+
+* Add slot for par_names to adfit objects
+
+* Add method `as.data.frame` for class `adfit`
+
+* Improved and expanded testing via continuous integration
+
+* Print the ADMB command to console when it fails to run properly to help user diagnose issues
+
+* Improve console output for RStudio users. It will now print at conclusion of parallel runs.
+
 * Fix bugs in model names for MacOS (use ./model insted of model internally)
+
 * Fix small bug with mceval=TRUE for newest verison of stock synthesis
 
 * Fix `sample_tmb` to work again for short-term use